Refactor Search components for improved loading state management
- Integrated `Suspense` and `withFallback` into `SearchButton` and `SearchEditor` components to enhance loading experiences.
- Removed redundant `Data` and `Pending` components to streamline the structure and improve maintainability.
- Updated various components to utilize colocated fallback UIs, ensuring consistent loading states across the application.
- Enhanced user interaction by adding `data-action` attributes for better accessibility and feedback.
