+++ apps/app/src/app/@seller/draft/DraftEditPage/DraftEditor/patch/AgePatch.tsx
@@ -10,14 +10,14 @@ import type { Rating } from "@zbav-se.me/ui/rating";
 import type { FC } from "react";
 import { AgeSelection } from "~/app/@common/age/ui/AgeSelection";
 import { SaveContainer } from "~/app/@common/container/ui/SaveContainer";
-import type { Data } from "../Data";
+import type { DraftEditor } from "../DraftEditor";
 import { EditAction } from "../EditAction";
 
 export namespace AgePatch {
 	export interface Props extends TitleContainer.Props {
 		draft: tDraft;
 		onCancel(): void;
-		onView(view: Data.View): void;
+		onView(view: DraftEditor.View): void;
 	}
 }
