+++ apps/app/src/app/@seller/draft/DraftEditPage/DraftEditor/patch/CategoryPatch.tsx
@@ -12,7 +12,7 @@ import { useAppForm } from "@zbav-se.me/ui/form";
 import type { FC } from "react";
 import { SaveContainer } from "~/app/@common/container/ui/SaveContainer";
 import { CategorySelect } from "~/app/@session/category/ui/CategorySelect/CategorySelect";
-import type { Data } from "../Data";
+import type { DraftEditor } from "../DraftEditor";
 import { EditAction } from "../EditAction";
 
 const CategorySchema = zListingCreate.pick({
@@ -23,7 +23,7 @@ export namespace CategoryPatch {
 	export interface Props extends TitleContainer.Props {
 		draft: tDraft;
 		onCancel(): void;
-		onView(view: Data.View): void;
+		onView(view: DraftEditor.View): void;
 	}
 }
