+++ apps/app/src/app/@seller/draft/DraftEditPage/DraftEditor/patch/PriceTypePatch.tsx
@@ -12,7 +12,7 @@ import { useAppForm } from "@zbav-se.me/ui/form";
 import type { FC } from "react";
 import { SaveContainer } from "~/app/@common/container/ui/SaveContainer";
 import { PriceTypeSelect } from "~/app/@common/price-type/ui/PriceTypeSelect";
-import type { Data } from "../Data";
+import type { DraftEditor } from "../DraftEditor";
 import { EditAction } from "../EditAction";
 
 const PriceTypeSchema = zListingCreate.pick({
@@ -23,7 +23,7 @@ export namespace PriceTypePatch {
 	export interface Props extends TitleContainer.Props {
 		draft: tDraft;
 		onCancel(): void;
-		onView(view: Data.View): void;
+		onView(view: DraftEditor.View): void;
 	}
 }
